tests/robustness: Separate describe model functions to dedicated file
